df32

+++ df32/routes/web.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\MovieController as M;
+use App\Http\Controllers\HomeController as H;
 
 /*
 |--------------------------------------------------------------------------
@@ -14,20 +15,16 @@ use App\Http\Controllers\MovieController as M;
 |
 */
 
-Route::get('/', function () {
-    return view('welcome');
-});
-
 Auth::routes();
-
-Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+Route::get('/', [H::class, 'homeList'])->name('home')->middleware('gate:home');
+Route::post('/comment/{movie}', [H::class, 'addComment'])->name('comment')->middleware('gate:user');
 
 Route::prefix('movie')->name('m_')->group(function () {
-    Route::get('/', [M::class, 'index'])->name('index');
-    Route::get('/create', [M::class, 'create'])->name('create');
-    Route::post('/create', [M::class, 'store'])->name('store');
-    Route::get('/show/{movie}', [M::class, 'show'])->name('show');
-    Route::delete('/delete/{movie}', [M::class, 'destroy'])->name('delete');
-    Route::get('/edit/{movie}', [M::class, 'edit'])->name('edit');
-    Route::put('/edit/{movie}', [M::class, 'update'])->name('update');
+    Route::get('/', [M::class, 'index'])->name('index')->middleware('gate:user');
+    Route::get('/create', [M::class, 'create'])->name('create')->middleware('gate:admin');
+    Route::post('/create', [M::class, 'store'])->name('store')->middleware('gate:admin');
+    Route::get('/show/{movie}', [M::class, 'show'])->name('show')->middleware('gate:user');
+    Route::delete('/delete/{movie}', [M::class, 'destroy'])->name('delete')->middleware('gate:admin');
+    Route::get('/edit/{movie}', [M::class, 'edit'])->name('edit')->middleware('gate:admin');
+    Route::put('/edit/{movie}', [M::class, 'update'])->name('update')->middleware('gate:admin');
 });
